% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_ablines}
\alias{show_ablines}
\title{Show geom_abline() graphs}
\usage{
show_ablines()
}
\value{
gg_penguins_abline, gg_penguins_smooth
}
\description{
Show geom_abline() graphs
}
\examples{
# not run
show_ablines()
}
