% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_area_bins}
\alias{show_area_bins}
\title{Show geom_area() graphs}
\usage{
show_area_bins()
}
\value{
gg_area_b15, gg_area_b30 , gg_area_b45, gg_area_bw1, gg_area_bw2,
gg_area_bw3
}
\description{
Show geom_area() graphs
}
\examples{
# not run
show_area_bins()
}
