% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_compare_binwidth}
\alias{show_compare_binwidth}
\title{Show bindwidth comparisons}
\usage{
show_compare_binwidth()
}
\value{
gg_dotplot, gg_freqpoly, gg_histogram
}
\description{
Show bindwidth comparisons
}
\examples{
# not run
show_compare_binwidth()
}
