% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_compare_dots}
\alias{show_compare_dots}
\title{Show geom_dotplot() graphs}
\usage{
show_compare_dots()
}
\value{
dotdensity, histodot
}
\description{
Show geom_dotplot() graphs
}
\examples{
# not run
show_compare_dots()
}
