% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_compare_qq}
\alias{show_compare_qq}
\title{Show geom_qq_line() &  geom_qq_line() graphs}
\usage{
show_compare_qq()
}
\value{
qq_rnorm, qq_rpois, qq_weibull, qq_cauchy
}
\description{
Show geom_qq_line() &  geom_qq_line() graphs
}
\examples{
# not run
show_compare_qq()
}
