% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_densities}
\alias{show_densities}
\title{Show geom_density() graphs}
\usage{
show_densities()
}
\value{
gaussian, biweight, cosine, optcosine
}
\description{
Show geom_density() graphs
}
\examples{
# not run
show_densities()
}
