% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_dotplot_bin_methods}
\alias{show_dotplot_bin_methods}
\title{Show geom_dotplot() binning methods}
\usage{
show_dotplot_bin_methods()
}
\value{
cyan, yellow, both_dotplots
}
\description{
Show geom_dotplot() binning methods
}
\examples{
# not run
show_dotplot_bin_methods()
}
