% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_expand_limits}
\alias{show_expand_limits}
\title{Show expand_limits() graphs}
\usage{
show_expand_limits()
}
\value{
gg_exp_lmt, gg_blnk, gg_a
}
\description{
Show expand_limits() graphs
}
\examples{
# not run
show_expand_limits()
}
