% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_geom_blanks}
\alias{show_geom_blanks}
\title{Show geom_blank() graphs}
\usage{
show_geom_blanks()
}
\value{
gg_a, gg_blnk
}
\description{
Show geom_blank() graphs
}
\examples{
# not run
show_geom_blanks()
}
