% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_geom_curve}
\alias{show_geom_curve}
\title{Show geom_curve() graphs}
\usage{
show_geom_curve()
}
\value{
gg_point, gg_curve
}
\description{
Show geom_curve() graphs
}
\examples{
# not run
show_geom_curve()
}
