% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_geom_path}
\alias{show_geom_path}
\title{Show geom_path() graphs}
\usage{
show_geom_path()
}
\value{
gg_pth_butt_bevel, gg_pth_round_mitre
}
\description{
Show geom_path() graphs
}
\examples{
# not run
show_geom_path()
}
