% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_geom_poly}
\alias{show_geom_poly}
\title{Show geom_poly() graphs}
\usage{
show_geom_poly()
}
\value{
gg_poly1, gg_poly2
}
\description{
Show geom_poly() graphs
}
\examples{
# not run
show_geom_poly()
}
