% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_geom_rect}
\alias{show_geom_rect}
\title{Show geom_rect() graphs}
\usage{
show_geom_rect()
}
\value{
gg_point, gg_curve, gg_rect, gg_point_rect
}
\description{
Show geom_rect() graphs
}
\examples{
# not run
show_geom_rect()
}
