% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_geom_ribbon}
\alias{show_geom_ribbon}
\title{Show geom_ribbon() graphs}
\usage{
show_geom_ribbon()
}
\value{
gg_rib1, gg_rib2
}
\description{
Show geom_ribbon() graphs
}
\examples{
# not run
show_geom_ribbon()
}
