% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_lines}
\alias{show_lines}
\title{Show geom_abline(), geom_hline(), geom_vline(), and geom_segment() graphs}
\usage{
show_lines()
}
\value{
gg_abline, gg_hline, gg_vline, gg_segment
}
\description{
Show geom_abline(), geom_hline(), geom_vline(), and geom_segment() graphs
}
\examples{
# not run
show_lines()
}
