% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_rep_spokes}
\alias{show_rep_spokes}
\title{Show geom_spoke() graphs (rep)}
\usage{
show_rep_spokes()
}
\value{
gg_spoke_1, gg_spoke_2, gg_spoke_3, gg_spoke_4
}
\description{
Show geom_spoke() graphs (rep)
}
\examples{
# not run
show_rep_spokes()
}
