% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp2_p1.R
\name{show_rnorm_spokes}
\alias{show_rnorm_spokes}
\title{Show geom_spoke() graphs (rnorm)}
\usage{
show_rnorm_spokes()
}
\value{
gg_spoke_a, gg_spoke_b, gg_spoke_c, gg_spoke_d
}
\description{
Show geom_spoke() graphs (rnorm)
}
\examples{
# not run
show_rnorm_spokes()
}
