% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_info.R
\name{vector_info}
\alias{vector_info}
\title{Get vector info}
\usage{
vector_info(x)
}
\arguments{
\item{x}{any vector (atomic or S3)}
}
\value{
a list of glue::glue() statements
}
\description{
vector_info(x) will return outputs from the following functions:
mode(x), class(x), typeof(x), attributes(names(x)), attributes(x).
}
\examples{
# not run
require(purrr)
log_vec <- c(TRUE, FALSE, TRUE, FALSE)
int_vec <- seq(1:4)
dbl_vec <- 1.1:4.1
chr_vec <- c("a", "b", "c", "d")
dat_vec <- c(Sys.Date(),
  Sys.Date() + 1,
  Sys.Date() + 2,
  Sys.Date() + 3)
dtt_vec <- c(Sys.time(),
  Sys.time() + 86400,
  Sys.time() + (2 * 86400),
  Sys.time() + (3 * 86400))
dft_vec <- c(
  difftime(
    time1 = Sys.time(),
    time2 = Sys.time() + (2 * 86400),
    units = "days"
  ),
  difftime(
    time1 = Sys.time(),
    time2 = Sys.time() + (3 * 86400),
    units = "days"
  ),
  difftime(
    time1 = Sys.time(),
    time2 = Sys.time() + (4 * 86400),
    units = "days"
  ),
  difftime(
    time1 = Sys.time(),
    time2 = Sys.time() + (5 * 86400),
    units = "days"
  )
)
fct_vec <- factor(
  x = c("never", "rarely", "often", "always"),
  levels = c("never", "rarely", "often", "always"),
  ordered = TRUE
)
 # list
vct_list <- list(
  'log_vec' = log_vec,
  'int_vec' = int_vec,
  'dbl_vec' = dbl_vec,
  'chr_vec' = chr_vec,

  'dat_vec' = dat_vec,
  'dtt_vec' = dtt_vec,
  'dft_vec' = dft_vec,

  'fct_vec' = fct_vec
)
purrr::map(.x = vct_list, .f = vector_info)
}
